+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. For commit guidelines, please refer to [Standard Version](https://github.com/conventional-changelog/standard-version).
 
+## v0.7.5-beta
+
+ **Notes**:
+ - logging uses localtime, optional UTC config added
+ - generated config example now includes defaults https://github.com/gtsteffaniak/filebrowser/issues/590
+
 ## v0.7.4-beta
 
  **Notes**: